Update podspec Update README

+++ FMMoveTable/Classes/FMViewController.m
@@ -74,7 +74,7 @@ static NSString *sCellIdentifier;
 
 - (NSString *)tableView:(UITableView *)tableView titleForHeaderInSection:(NSInteger)section
 {
-	return [NSString stringWithFormat:@"Section %i", section];
+	return [NSString stringWithFormat:@"Section %li", (long)section];
 }
